adding a contact page
the metadata for the app seems to contain this url and the certification
fails with the following error message.

Notes to publisher	10.1.2.7 Functionality
There was an issue with URL links included in the product metadata. The links did not resolve to a functional website.
Affected Link: Support Contact: https://www.youmacro.com/contact/